change to path import and running direct py script
