+++ makefile
@@ -5,8 +5,8 @@ python = python3.8
 make:
 	${run_carla} -nosound -RenderOffScreen & \
 	sleep 3 && \
-		${python} main.py speed_sensor run_speed_sensor.py & \
-		${python} main.py drive run_drive.py
+		${python} speed_sensor/run_speed_sensor.py & \
+		${python} drive/run_drive.py
 
 kill:
 	pgrep -f ${run_carla} | kill
